#283 Separating Caching, Core and Session component builders in two parts, one containing "With" and one "Without" implementations.
