+++ src/MyTested.AspNetCore.Mvc.Caching/ComponentBuilderCachingWithoutExtensions.cs
@@ -10,26 +10,8 @@
     /// <summary>
     /// Contains <see cref="Microsoft.Extensions.Caching.Memory.IMemoryCache"/> extension methods for <see cref="IBaseTestBuilderWithComponentBuilder{TBuilder}"/>.
     /// </summary>
-    public static class ComponentBuilderCachingExtensions
+    public static class ComponentBuilderCachingWithoutExtensions
     {
-        /// <summary>
-        /// Sets initial values to the <see cref="Microsoft.Extensions.Caching.Memory.IMemoryCache"/> service.
-        /// </summary>
-        /// <typeparam name="TBuilder">Class representing ASP.NET Core MVC test builder.</typeparam>
-        /// <param name="builder">Instance of <see cref="IBaseTestBuilderWithComponentBuilder{TBuilder}"/> type.</param>
-        /// <param name="memoryCacheBuilder">Action setting the <see cref="Microsoft.Extensions.Caching.Memory.IMemoryCache"/> values by using <see cref="IWithMemoryCacheBuilder"/>.</param>
-        /// <returns>The same component builder.</returns>
-        public static TBuilder WithMemoryCache<TBuilder>(
-            this IBaseTestBuilderWithComponentBuilder<TBuilder> builder,
-            Action<IWithMemoryCacheBuilder> memoryCacheBuilder)
-            where TBuilder : IBaseTestBuilder
-        {
-            var actualBuilder = (BaseTestBuilderWithComponentBuilder<TBuilder>)builder;
-
-            memoryCacheBuilder(new WithMemoryCacheBuilder(actualBuilder.TestContext.HttpContext.RequestServices));
-
-            return actualBuilder.Builder;
-        }
 
         /// <summary>
         /// Clear all entities from <see cref="Microsoft.Extensions.Caching.Memory.IMemoryCache"/> service.
@@ -50,7 +32,7 @@
         /// <param name="key">Key of the entity that will be removed.</param>
         /// <returns>The same component builder.</returns>
         public static TBuilder WithoutMemoryCache<TBuilder>(
-            this IBaseTestBuilderWithComponentBuilder<TBuilder> builder, 
+            this IBaseTestBuilderWithComponentBuilder<TBuilder> builder,
             object key)
             where TBuilder : IBaseTestBuilder
             => builder.WithoutMemoryCache(cache => cache.WithoutEntry(key));
